chore: move ComponentExecution to execution.go
